Web: Files: refactoring FilesRowContent, withContentActions
